Organize package under src/ layout
