+++ examples/04_error_handling.py
@@ -6,7 +6,7 @@ from datetime import datetime, timedelta
 # Add the parent directory to the Python path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-from bcra_connector import BCRAConnector, BCRAApiError
+from src.bcra_connector import BCRAConnector
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
 logger = logging.getLogger(__name__)
